requirements.txt
fastapi==0.110.3
uvicorn==0.23.1

tensorflow==2.11.0              # Last officially supported for Python 3.10
torch==2.1.0                    # Compatible with Python 3.10
torchvision==0.16.0

scikit-learn==1.3.2
scipy==1.11.4
numpy==1.26.4
pandas==2.2.2

opencv-python==4.10.0.84
Pillow==9.5.0
matplotlib==3.7.3
imageio==2.34.1
tifffile==2023.7.10
scikit-image==0.22.0

python-multipart==0.0.6
python-dotenv==1.0.1
requests==2.31.0

pydantic==2.3.0
pydantic-settings==2.1.0


PyYAML==6.0.1

pytest==7.4.4
httpx==0.27.0


rich==13.7.0

tensorboard==2.11.2
protobuf==3.19.6

